refactor: flatten pkg/routefinder/rfclient to pkg/rfclient
The route-finder client lived two levels deep at pkg/routefinder/rfclient
while every other discovery client sits at the package root
(pkg/utclient, pkg/transport/tpdclient). The doubled directory level
was a remnant of an earlier layout and added nothing — pkg/routefinder
contained only the rfclient subdir.

Move pkg/routefinder/rfclient → pkg/rfclient (matches pkg/utclient).
Remove now-empty pkg/routefinder/. Service implementation at
pkg/route-finder/ stays (matches pkg/transport-discovery,
pkg/service-discovery, pkg/address-resolver service-package convention).
